+++ packages/ui/CHANGELOG.md
@@ -6,6 +6,8 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
+- `ui/ProgressBar` to support additional `orientation` type of `'radial'`
+- `ui/ProgressBar` public class name `radial`
 - `ui/ViewManager` events `onTransition` and `onWillTransition` payload members `index` and `previousIndex`
 
 ### Fixed